requirements.txt
Flask==3.0.3
Werkzeug==3.1.3
Jinja2==3.1.6
itsdangerous==2.2.0
blinker==1.9.0
SQLAlchemy==2.0.43
requests==2.32.5
python-dotenv==1.1.1
pytest==8.4.2
pytest-cov==7.0.0
pytest-json-report==1.5.0
gunicorn==23.0.0
